Split examples from VIA core to separate file
