+++ VIA/__init__.py
@@ -1,1 +1,2 @@
 from . import core
+from . import examples
